Make tab functions to take size_t instead of unsigned int as size parameter

+++ tab/ft_rev_int_tab.c
@@ -6,16 +6,16 @@
 /*   By: chlimous <chlimous@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/03 01:42:04 by chlimous          #+#    #+#             */
-/*   Updated: 2024/01/04 01:15:03 by chlimous         ###   ########.fr       */
+/*   Updated: 2024/01/06 00:14:03 by chlimous         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_rev_int_tab(int *tab, unsigned int size)
+void	ft_rev_int_tab(int *tab, size_t size)
 {
-	unsigned int	i;
-	int				temp;
+	size_t	i;
+	int	temp;
 
 	i = 0;
 	while (i < size / 2)
