+++ tab/ft_rev_long_tab.c
@@ -6,16 +6,16 @@
 /*   By: chlimous <chlimous@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/04 01:06:10 by chlimous          #+#    #+#             */
-/*   Updated: 2024/01/04 01:14:47 by chlimous         ###   ########.fr       */
+/*   Updated: 2024/01/06 00:14:21 by chlimous         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_rev_long_tab(long *tab, unsigned int size)
+void	ft_rev_long_tab(long *tab, size_t size)
 {
-	unsigned int	i;
-	long			temp;
+	size_t	i;
+	long	temp;
 
 	i = 0;
 	while (i < size / 2)
